implement entry renaming in engine
