+++ reductstore/src/storage/bucket.rs
@@ -1,10 +1,16 @@
-// Copyright 2023 ReductSoftware UG
+// Copyright 2023-2024 ReductSoftware UG
 // Licensed under the Business Source License 1.1
+
+mod quotas;
+mod settings;
 
 use crate::core::thread_pool::{unique, TaskHandle};
 use crate::core::weak::Weak;
 pub use crate::storage::block_manager::RecordRx;
 pub use crate::storage::block_manager::RecordTx;
+use crate::storage::bucket::settings::{
+    DEFAULT_MAX_BLOCK_SIZE, DEFAULT_MAX_RECORDS, SETTINGS_NAME,
+};
 use crate::storage::entry::{Entry, EntrySettings, RecordReader, WriteRecordContent};
 use crate::storage::file_cache::FILE_CACHE;
 use crate::storage::proto::BucketSettings as ProtoBucketSettings;
@@ -14,46 +20,12 @@ use prost::Message;
 use reduct_base::error::ReductError;
 use reduct_base::msg::bucket_api::{BucketInfo, BucketSettings, FullBucketInfo, QuotaType};
 use reduct_base::msg::entry_api::EntryInfo;
-use reduct_base::{bad_request, conflict, internal_server_error, Labels};
+use reduct_base::{bad_request, conflict, internal_server_error, not_found, Labels};
 use std::collections::BTreeMap;
 use std::io::Write;
 use std::path::PathBuf;
 use std::sync::atomic::AtomicBool;
 use std::sync::{Arc, RwLock};
-
-const DEFAULT_MAX_RECORDS: u64 = 1024;
-const DEFAULT_MAX_BLOCK_SIZE: u64 = 64000000;
-const SETTINGS_NAME: &str = "bucket.settings";
-
-impl From<BucketSettings> for ProtoBucketSettings {
-    fn from(settings: BucketSettings) -> Self {
-        ProtoBucketSettings {
-            quota_size: settings.quota_size,
-            quota_type: if let Some(quota_type) = settings.quota_type {
-                Some(quota_type as i32)
-            } else {
-                None
-            },
-            max_block_records: settings.max_block_records,
-            max_block_size: settings.max_block_size,
-        }
-    }
-}
-
-impl Into<BucketSettings> for ProtoBucketSettings {
-    fn into(self) -> BucketSettings {
-        BucketSettings {
-            quota_size: self.quota_size,
-            quota_type: if let Some(quota_type) = self.quota_type {
-                Some(QuotaType::from(quota_type))
-            } else {
-                None
-            },
-            max_block_records: self.max_block_records,
-            max_block_size: self.max_block_size,
-        }
-    }
-}
 
 /// Bucket is a single storage bucket.
 pub(crate) struct Bucket {
@@ -148,16 +120,6 @@ impl Bucket {
             settings: RwLock::new(settings),
             is_provisioned: AtomicBool::new(false),
         })
-    }
-
-    /// Default settings for a new bucket bucket
-    pub fn defaults() -> BucketSettings {
-        BucketSettings {
-            max_block_size: Some(DEFAULT_MAX_BLOCK_SIZE),
-            quota_type: Some(QuotaType::NONE),
-            quota_size: Some(0),
-            max_block_records: Some(DEFAULT_MAX_RECORDS),
-        }
     }
 
     /// Get or create an entry in the bucket
@@ -290,6 +252,50 @@ impl Bucket {
         }
     }
 
+    pub fn rename_entry(&self, old_name: &str, new_name: &str) -> Result<(), ReductError> {
+        let old_path = self.path.join(old_name);
+        let new_path = self.path.join(new_name);
+        let bucket_name = self.name.clone();
+        let mut entries = self.entries.clone();
+        let old_name = old_name.to_string();
+
+        unique(&self.task_group(), "rename entry", move || {
+            if (entries.write().unwrap().remove(&old_name).is_none()) {
+                return Err(not_found!(
+                    "Entry '{}' not found in bucket '{}'",
+                    old_name,
+                    bucket_name
+                ));
+            }
+
+            std::fs::rename(&old_path, &new_path)?;
+            Ok(())
+        })
+        .wait()?;
+
+        let entry = Entry::restore(
+            self.path.join(new_name.clone()),
+            EntrySettings {
+                max_block_size: self
+                    .settings
+                    .read()?
+                    .max_block_size
+                    .unwrap_or(DEFAULT_MAX_BLOCK_SIZE),
+                max_block_records: self
+                    .settings
+                    .read()?
+                    .max_block_records
+                    .unwrap_or(DEFAULT_MAX_RECORDS),
+            },
+        )
+        .wait()?;
+
+        self.entries
+            .write()?
+            .insert(new_name.to_string(), Arc::new(entry));
+        Ok(())
+    }
+
     /// Remove entry from the bucket
     ///
     /// # Arguments
@@ -323,83 +329,6 @@ impl Bucket {
         })
     }
 
-    fn keep_quota_for(&self, content_size: usize) -> Result<(), ReductError> {
-        let settings = self.settings.read()?;
-        let quota_size = settings.quota_size.unwrap_or(0);
-        match settings.quota_type.clone().unwrap_or(QuotaType::NONE) {
-            QuotaType::NONE => Ok(()),
-            QuotaType::FIFO => self.remove_oldest_block(content_size, quota_size),
-            QuotaType::HARD => {
-                if self.info()?.info.size + content_size as u64 > quota_size {
-                    Err(bad_request!("Quota of '{}' exceeded", self.name()))
-                } else {
-                    Ok(())
-                }
-            }
-        }
-    }
-
-    fn remove_oldest_block(&self, content_size: usize, quota_size: u64) -> Result<(), ReductError> {
-        let mut size = self.info()?.info.size + content_size as u64;
-        while size > quota_size {
-            debug!(
-                "Need more space. Remove an oldest block from bucket '{}'",
-                self.name()
-            );
-
-            let mut candidates: Vec<(u64, &Entry)> = vec![];
-            let entries = self.entries.read()?;
-            for (_, entry) in entries.iter() {
-                let info = entry.info()?;
-                candidates.push((info.oldest_record, entry));
-            }
-            candidates.sort_by_key(|entry| entry.0);
-
-            let candidates = candidates
-                .iter()
-                .map(|(_, entry)| entry.name().to_string())
-                .collect::<Vec<String>>();
-
-            let mut success = false;
-            for name in candidates {
-                debug!("Remove an oldest block from entry '{}'", name);
-                match entries.get(&name).unwrap().try_remove_oldest_block().wait() {
-                    Ok(_) => {
-                        success = true;
-                        break;
-                    }
-                    Err(e) => {
-                        debug!("Failed to remove oldest block from entry '{}': {}", name, e);
-                    }
-                }
-            }
-
-            if !success {
-                return Err(internal_server_error!(
-                    "Failed to keep quota of '{}'",
-                    self.name()
-                ));
-            }
-
-            size = self.info()?.info.size + content_size as u64;
-        }
-
-        // Remove empty entries
-        let mut entries = self.entries.write()?;
-        let mut names_to_remove = vec![];
-        for (name, entry) in entries.iter() {
-            if entry.info()?.record_count != 0 {
-                continue;
-            }
-            names_to_remove.push(name.clone());
-        }
-
-        for name in names_to_remove {
-            entries.remove(&name);
-        }
-        Ok(())
-    }
-
     /// Sync all entries to the file system
     pub fn sync_fs(&self) -> TaskHandle<Result<(), ReductError>> {
         let entries = self.entries.clone();
@@ -417,37 +346,6 @@ impl Bucket {
         &self.name
     }
 
-    pub fn settings(&self) -> BucketSettings {
-        self.settings.read().unwrap().clone()
-    }
-
-    pub fn set_settings(&self, settings: BucketSettings) -> TaskHandle<Result<(), ReductError>> {
-        if self
-            .is_provisioned
-            .load(std::sync::atomic::Ordering::Relaxed)
-        {
-            return Err(conflict!(
-                "Can't change settings of provisioned bucket '{}'",
-                self.name()
-            ))
-            .into();
-        }
-
-        {
-            let mut my_settings = self.settings.write().unwrap();
-            let entries = self.entries.write().unwrap();
-
-            *my_settings = Self::fill_settings(settings, my_settings.clone());
-            for entry in entries.values() {
-                entry.set_settings(EntrySettings {
-                    max_block_size: my_settings.max_block_size.unwrap(),
-                    max_block_records: my_settings.max_block_records.unwrap(),
-                });
-            }
-        }
-        self.save_settings()
-    }
-
     /// Mark bucket as provisioned to protect
     pub fn set_provisioned(&self, provisioned: bool) {
         self.is_provisioned
@@ -458,40 +356,6 @@ impl Bucket {
     pub fn is_provisioned(&self) -> bool {
         self.is_provisioned
             .load(std::sync::atomic::Ordering::Relaxed)
-    }
-
-    /// Fill in missing settings with defaults
-    fn fill_settings(settings: BucketSettings, default: BucketSettings) -> BucketSettings {
-        let mut settings = settings;
-        if settings.max_block_size.is_none() {
-            settings.max_block_size = default.max_block_size;
-        }
-        if settings.quota_type.is_none() {
-            settings.quota_type = default.quota_type;
-        }
-        if settings.quota_size.is_none() {
-            settings.quota_size = default.quota_size;
-        }
-        if settings.max_block_records.is_none() {
-            settings.max_block_records = default.max_block_records;
-        }
-        settings
-    }
-
-    fn save_settings(&self) -> TaskHandle<Result<(), ReductError>> {
-        let settings = self.settings.read().unwrap().clone();
-        let path = self.path.join(SETTINGS_NAME);
-
-        unique(&self.task_group(), "save settings", move || {
-            let mut buf = BytesMut::new();
-            ProtoBucketSettings::from(settings)
-                .encode(&mut buf)
-                .map_err(|e| internal_server_error!("Failed to encode bucket settings: {}", e))?;
-
-            let mut file = std::fs::File::create(path)?;
-            file.write(&buf)?;
-            Ok(())
-        })
     }
 
     fn task_group(&self) -> String {
@@ -516,165 +380,91 @@ mod tests {
     use rstest::{fixture, rstest};
     use tempfile::tempdir;
 
-    #[rstest]
-    fn test_keep_settings_persistent(settings: BucketSettings, bucket: Bucket) {
-        assert_eq!(bucket.settings(), settings);
+    mod remove_entry {
+        use super::*;
 
-        let bucket = Bucket::restore(bucket.path.clone()).unwrap();
-        assert_eq!(bucket.name(), "test");
-        assert_eq!(bucket.settings(), settings);
-    }
+        #[rstest]
+        #[tokio::test]
+        async fn test_remove_entry(mut bucket: Bucket) {
+            write(&mut bucket, "test-1", 1, b"test").await.unwrap();
 
-    #[rstest]
-    fn test_fill_default_settings() {
-        let settings = BucketSettings {
-            max_block_size: None,
-            quota_type: None,
-            quota_size: None,
-            max_block_records: None,
-        };
+            bucket.remove_entry("test-1").await.unwrap();
+            assert_eq!(
+                bucket.get_entry("test-1").err(),
+                Some(ReductError::not_found(
+                    "Entry 'test-1' not found in bucket 'test'"
+                ))
+            );
+        }
 
-        let default_settings = Bucket::defaults();
-        let filled_settings = Bucket::fill_settings(settings, default_settings.clone());
-        assert_eq!(filled_settings, default_settings);
-    }
-
-    #[rstest]
-    fn test_set_settings_partially(settings: BucketSettings, bucket: Bucket) {
-        let new_settings = BucketSettings {
-            max_block_size: Some(100),
-            quota_type: None,
-            quota_size: None,
-            max_block_records: None,
-        };
-
-        bucket.set_settings(new_settings).wait().unwrap();
-        assert_eq!(bucket.settings().max_block_size.unwrap(), 100);
-        assert_eq!(bucket.settings().quota_type, settings.quota_type);
-        assert_eq!(bucket.settings().quota_size, settings.quota_size);
-        assert_eq!(
-            bucket.settings().max_block_records,
-            settings.max_block_records
-        );
-    }
-
-    #[rstest]
-    fn test_apply_settings_to_entries(settings: BucketSettings, bucket: Bucket) {
-        bucket.get_or_create_entry("entry-1").unwrap();
-        bucket.get_or_create_entry("entry-2").unwrap();
-
-        let mut new_settings = settings.clone();
-        new_settings.max_block_size = Some(200);
-        new_settings.max_block_records = Some(200);
-        bucket.set_settings(new_settings).wait().unwrap();
-
-        for entry in bucket.entries.read().unwrap().values() {
-            assert_eq!(entry.settings().max_block_size, 200);
-            assert_eq!(entry.settings().max_block_records, 200);
+        #[rstest]
+        #[tokio::test]
+        async fn test_remove_entry_not_found(bucket: Bucket) {
+            assert_eq!(
+                bucket.remove_entry("test-1").await.err(),
+                Some(ReductError::not_found(
+                    "Entry 'test-1' not found in bucket 'test'"
+                ))
+            );
         }
     }
 
-    #[rstest]
-    #[tokio::test]
-    async fn test_fifo_quota_keeping(path: PathBuf) {
-        let mut bucket = bucket(
-            BucketSettings {
-                max_block_size: Some(20),
-                quota_type: Some(QuotaType::FIFO),
-                quota_size: Some(100),
-                max_block_records: Some(100),
-            },
-            path,
-        );
+    mod rename_entry {
+        use super::*;
 
-        let blob: &[u8] = &[0u8; 40];
+        #[rstest]
+        #[tokio::test]
+        async fn test_rename_entry(mut bucket: Bucket) {
+            write(&mut bucket, "test-1", 1, b"test").await.unwrap();
 
-        write(&mut bucket, "test-1", 0, blob).await.unwrap();
-        assert_eq!(bucket.info().unwrap().info.size, 44);
+            bucket.rename_entry("test-1", "test-2").unwrap();
+            assert_eq!(
+                bucket.get_entry("test-1").err(),
+                Some(ReductError::not_found(
+                    "Entry 'test-1' not found in bucket 'test'"
+                ))
+            );
+            assert_eq!(
+                bucket
+                    .get_entry("test-2")
+                    .unwrap()
+                    .upgrade()
+                    .unwrap()
+                    .name(),
+                "test-2"
+            );
+        }
 
-        write(&mut bucket, "test-2", 1, blob).await.unwrap();
-        assert_eq!(bucket.info().unwrap().info.size, 91);
+        #[rstest]
+        #[tokio::test]
+        async fn test_rename_entry_not_found(bucket: Bucket) {
+            assert_eq!(
+                bucket.rename_entry("test-1", "test-2").err(),
+                Some(not_found!("Entry 'test-1' not found in bucket 'test'"))
+            );
+        }
 
-        write(&mut bucket, "test-3", 2, blob).await.unwrap();
-        assert_eq!(bucket.info().unwrap().info.size, 94);
+        #[rstest]
+        #[tokio::test]
+        async fn test_rename_entry_persisted(mut bucket: Bucket) {
+            write(&mut bucket, "test-1", 1, b"test").await.unwrap();
+            bucket.sync_fs().await.unwrap();
+            bucket.rename_entry("test-1", "test-2").unwrap();
 
-        assert_eq!(
-            read(&mut bucket, "test-1", 0).await.err(),
-            Some(ReductError::not_found(
-                "Entry 'test-1' not found in bucket 'test'"
-            ))
-        );
-    }
+            let bucket = Bucket::restore(bucket.path.clone()).unwrap();
+            assert_eq!(
+                bucket.get_entry("test-1").err(),
+                Some(ReductError::not_found(
+                    "Entry 'test-1' not found in bucket 'test'"
+                ))
+            );
 
-    #[rstest]
-    #[tokio::test]
-    async fn test_hard_quota_keeping(path: PathBuf) {
-        let mut bucket = bucket(
-            BucketSettings {
-                quota_type: Some(QuotaType::HARD),
-                quota_size: Some(100),
-                ..BucketSettings::default()
-            },
-            path,
-        );
-
-        let blob: &[u8] = &[0u8; 40];
-        write(&mut bucket, "test-1", 0, blob).await.unwrap();
-        write(&mut bucket, "test-2", 1, blob).await.unwrap();
-
-        let err = write(&mut bucket, "test-3", 2, blob).await.err().unwrap();
-        assert_eq!(err, ReductError::bad_request("Quota of 'test' exceeded"));
-    }
-
-    #[rstest]
-    #[tokio::test]
-    async fn test_blob_bigger_than_quota(path: PathBuf) {
-        let mut bucket = bucket(
-            BucketSettings {
-                max_block_size: Some(5),
-                quota_type: Some(QuotaType::FIFO),
-                quota_size: Some(10),
-                max_block_records: Some(100),
-            },
-            path,
-        );
-
-        write(&mut bucket, "test-1", 0, b"test").await.unwrap();
-        bucket.sync_fs().await.unwrap(); // we need to sync to get the correct size
-        assert_eq!(bucket.info().unwrap().info.size, 22);
-
-        let result = write(&mut bucket, "test-2", 1, b"0123456789___").await;
-        assert_eq!(
-            result.err(),
-            Some(ReductError::internal_server_error(
-                "Failed to keep quota of 'test'"
-            ))
-        );
-    }
-
-    #[rstest]
-    #[tokio::test]
-    async fn test_remove_entry(mut bucket: Bucket) {
-        write(&mut bucket, "test-1", 1, b"test").await.unwrap();
-
-        bucket.remove_entry("test-1").await.unwrap();
-        assert_eq!(
-            bucket.get_entry("test-1").err(),
-            Some(ReductError::not_found(
-                "Entry 'test-1' not found in bucket 'test'"
-            ))
-        );
-    }
-
-    #[rstest]
-    #[tokio::test]
-    async fn test_remove_entry_not_found(bucket: Bucket) {
-        assert_eq!(
-            bucket.remove_entry("test-1").await.err(),
-            Some(ReductError::not_found(
-                "Entry 'test-1' not found in bucket 'test'"
-            ))
-        );
+            let mut reader = bucket.begin_read("test-2", 1).await.unwrap();
+            assert_eq!(
+                reader.rx().recv().await.unwrap().unwrap(),
+                Bytes::from("test")
+            );
+        }
     }
 
     #[rstest]
@@ -696,7 +486,7 @@ mod tests {
         );
     }
 
-    async fn write(
+    pub async fn write(
         bucket: &mut Bucket,
         entry_name: &str,
         time: u64,
@@ -725,7 +515,7 @@ mod tests {
         Ok(())
     }
 
-    async fn read(
+    pub async fn read(
         bucket: &mut Bucket,
         entry_name: &str,
         time: u64,
@@ -736,7 +526,7 @@ mod tests {
     }
 
     #[fixture]
-    fn settings() -> BucketSettings {
+    pub fn settings() -> BucketSettings {
         BucketSettings {
             max_block_size: Some(100),
             quota_type: Some(QuotaType::FIFO),
@@ -746,17 +536,17 @@ mod tests {
     }
 
     #[fixture]
-    fn path() -> PathBuf {
+    pub fn path() -> PathBuf {
         tempdir().unwrap().into_path()
     }
 
     #[fixture]
-    fn bucket(settings: BucketSettings, path: PathBuf) -> Bucket {
+    pub fn bucket(settings: BucketSettings, path: PathBuf) -> Bucket {
         Bucket::new("test", &path, settings).unwrap()
     }
 
     #[fixture]
-    fn provisioned_bucket(settings: BucketSettings, path: PathBuf) -> Bucket {
+    pub fn provisioned_bucket(settings: BucketSettings, path: PathBuf) -> Bucket {
         let bucket = Bucket::new("test", &path, settings).unwrap();
         bucket.set_provisioned(true);
         bucket
